requirements.txt
# Core scientific computing
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.0.0

# For dimensionality reduction
umap-learn>=0.5.0

# For HTML report generation
jinja2>=3.0.0

# Additional utilities
pandas>=1.3.0
tqdm>=4.60.0
pyyaml>=6.0